# Source/Zybach.API/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Zybach",
    "version": "v1"
  },
  "paths": {
...
    "/api/chemigationPermits/{chemigationPermitNumber}/annualRecords": {
      "get": {
        "tags": [
          "ChemigationPermitAnnualRecord"
        ],
        "parameters": [
          {
            "name": "chemigationPermitNumber",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ChemigationPermitAnnualRecordDetailedDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ChemigationPermitAnnualRecordDetailedDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ChemigationPermitAnnualRecordDetailedDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/chemigationPermits/{chemigationPermitNumber}/getLatestRecordYear": {
      "get": {
        "tags": [
          "ChemigationPermitAnnualRecord"
        ],
        "parameters": [
          {
            "name": "chemigationPermitNumber",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ChemigationPermitAnnualRecordDetailedDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChemigationPermitAnnualRecordDetailedDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChemigationPermitAnnualRecordDetailedDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/chemigationPermits/{chemigationPermitNumber}/{recordYear}": {
      "get": {
        "tags": [
          "ChemigationPermitAnnualRecord"
        ],
        "parameters": [
          {
            "name": "chemigationPermitNumber",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "recordYear",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ChemigationPermitAnnualRecordDetailedDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChemigationPermitAnnualRecordDetailedDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChemigationPermitAnnualRecordDetailedDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/chemigationPermits/annualRecords/{chemigationPermitAnnualRecordID}": {
      "put": {
        "tags": [
          "ChemigationPermitAnnualRecord"
        ],
        "parameters": [
          {
            "name": "chemigationPermitAnnualRecordID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/ChemigationPermitAnnualRecordUpsertDto"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChemigationPermitAnnualRecordUpsertDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ChemigationPermitAnnualRecordUpsertDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ChemigationPermitAnnualRecordUpsertDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ChemigationPermitAnnualRecordDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChemigationPermitAnnualRecordDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChemigationPermitAnnualRecordDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/chemigationPermits/{chemigationPermitID}/annualRecords": {
      "post": {
        "tags": [
          "ChemigationPermitAnnualRecord"
        ],
        "parameters": [
          {
            "name": "chemigationPermitID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/ChemigationPermitAnnualRecordUpsertDto"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChemigationPermitAnnualRecordUpsertDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ChemigationPermitAnnualRecordUpsertDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ChemigationPermitAnnualRecordUpsertDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ChemigationPermitAnnualRecordDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChemigationPermitAnnualRecordDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChemigationPermitAnnualRecordDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/chemigationPermits/annualRecordsBulkCreate/{recordYear}": {
      "post": {
        "tags": [
          "ChemigationPermitAnnualRecord"
        ],
        "parameters": [
          {
            "name": "recordYear",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "text/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChemigationInjectionUnitTypeDto": {
        "type": "object",
        "properties": {
          "ChemigationInjectionUnitTypeID": {
            "type": "integer",
            "format": "int32"
          },
          "ChemigationInjectionUnitTypeName": {
            "type": "string",
            "nullable": true
          },
          "ChemigationInjectionUnitTypeDisplayName": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ChemigationPermitAnnualRecordApplicatorSimpleDto": {
        "type": "object",
        "properties": {
          "ChemigationPermitAnnualRecordApplicatorID": {
            "type": "integer",
            "format": "int32"
          },
          "ChemigationPermitAnnualRecordID": {
            "type": "integer",
            "format": "int32"
          },
          "ApplicatorName": {
            "type": "string",
            "nullable": true
          },
          "CertificationNumber": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "ExpirationYear": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "HomePhone": {
            "type": "string",
            "nullable": true
          },
          "MobilePhone": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ChemigationPermitAnnualRecordApplicatorUpsertDto": {
        "required": [
          "ApplicatorName",
          "CertificationNumber",
          "ExpirationYear"
        ],
        "type": "object",
        "properties": {
          "ChemigationPermitAnnualRecordApplicatorID": {
            "type": "integer",
            "format": "int32"
          },
          "ChemigationPermitAnnualRecordID": {
            "type": "integer",
            "format": "int32"
          },
          "ApplicatorName": {
            "type": "string"
          },
          "CertificationNumber": {
            "type": "integer",
            "format": "int32"
          },
          "ExpirationYear": {
            "type": "integer",
            "format": "int32"
          },
          "HomePhone": {
            "pattern": "^\\(?([0-9]{3})\\)?[-. ]?([0-9]{3})[-. ]?([0-9]{4})$",
            "type": "string",
            "nullable": true
          },
          "MobilePhone": {
            "pattern": "^\\(?([0-9]{3})\\)?[-. ]?([0-9]{3})[-. ]?([0-9]{4})$",
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ChemigationPermitAnnualRecordChemicalFormulationSimpleDto": {
        "type": "object",
        "properties": {
          "ChemicalFormulationName": {
            "type": "string",
            "nullable": true
          },
          "ChemicalUnitName": {
            "type": "string",
            "nullable": true
          },
          "ChemigationPermitAnnualRecordChemicalFormulationID": {
            "type": "integer",
            "format": "int32"
          },
          "ChemigationPermitAnnualRecordID": {
            "type": "integer",
            "format": "int32"
          },
          "ChemicalFormulationID": {
            "type": "integer",
            "format": "int32"
          },
          "ChemicalUnitID": {
            "type": "integer",
            "format": "int32"
          },
          "TotalApplied": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "AcresTreated": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
      "ChemigationPermitAnnualRecordChemicalFormulationUpsertDto": {
        "type": "object",
        "properties": {
          "ChemigationPermitAnnualRecordChemicalFormulationID": {
            "type": "integer",
            "format": "int32"
          },
          "ChemigationPermitAnnualRecordID": {
            "type": "integer",
            "format": "int32"
          },
          "ChemicalFormulationID": {
            "type": "integer",
            "format": "int32"
          },
          "ChemicalUnitID": {
            "type": "integer",
            "format": "int32"
          },
          "TotalApplied": {
            "maximum": 999999,
            "minimum": 0,
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "AcresTreated": {
            "maximum": 999999,
            "minimum": 0,
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
      "ChemigationPermitAnnualRecordDetailedDto": {
        "type": "object",
        "properties": {
          "ChemigationPermitAnnualRecordID": {
            "type": "integer",
            "format": "int32"
          },
          "ChemigationPermit": {
            "$ref": "#/components/schemas/ChemigationPermitDto"
          },
          "RecordYear": {
            "type": "integer",
            "format": "int32"
          },
          "ChemigationPermitAnnualRecordStatusID": {
            "type": "integer",
            "format": "int32"
          },
          "ChemigationPermitAnnualRecordStatusName": {
            "type": "string",
            "nullable": true
          },
          "PivotName": {
            "type": "string",
            "nullable": true
          },
          "ChemigationInjectionUnitTypeID": {
            "type": "integer",
            "format": "int32"
          },
          "ChemigationInjectionUnitTypeName": {
            "type": "string",
            "nullable": true
          },
          "ApplicantName": {
            "type": "string",
            "nullable": true
          },
          "ApplicantMailingAddress": {
            "type": "string",
            "nullable": true
          },
          "ApplicantCity": {
            "type": "string",
            "nullable": true
          },
          "ApplicantState": {
            "type": "string",
            "nullable": true
          },
          "ApplicantZipCode": {
            "type": "string",
            "nullable": true
          },
          "ApplicantPhone": {
            "type": "string",
            "nullable": true
          },
          "ApplicantMobilePhone": {
            "type": "string",
            "nullable": true
          },
          "DateReceived": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "DatePaid": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "ApplicantEmail": {
            "type": "string",
            "nullable": true
          },
          "ChemicalFormulations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChemigationPermitAnnualRecordChemicalFormulationSimpleDto"
            },
            "nullable": true
          },
          "Applicators": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChemigationPermitAnnualRecordApplicatorSimpleDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ChemigationPermitAnnualRecordDto": {
        "type": "object",
        "properties": {
          "ChemigationPermitAnnualRecordID": {
            "type": "integer",
            "format": "int32"
          },
          "ChemigationPermit": {
            "$ref": "#/components/schemas/ChemigationPermitDto"
          },
          "RecordYear": {
            "type": "integer",
            "format": "int32"
          },
          "ChemigationPermitAnnualRecordStatus": {
            "$ref": "#/components/schemas/ChemigationPermitAnnualRecordStatusDto"
          },
          "PivotName": {
            "type": "string",
            "nullable": true
          },
          "ChemigationInjectionUnitType": {
            "$ref": "#/components/schemas/ChemigationInjectionUnitTypeDto"
          },
          "ApplicantName": {
            "type": "string",
            "nullable": true
          },
          "ApplicantMailingAddress": {
            "type": "string",
            "nullable": true
          },
          "ApplicantCity": {
            "type": "string",
            "nullable": true
          },
          "ApplicantState": {
            "type": "string",
            "nullable": true
          },
          "ApplicantZipCode": {
            "type": "string",
            "nullable": true
          },
          "ApplicantPhone": {
            "type": "string",
            "nullable": true
          },
          "ApplicantMobilePhone": {
            "type": "string",
            "nullable": true
          },
          "DateReceived": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "DatePaid": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "ApplicantEmail": {
            "type": "string",
            "nullable": true
          },
          "NDEEAmount": {
            "type": "number",
            "format": "double",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ChemigationPermitAnnualRecordStatusDto": {
        "type": "object",
        "properties": {
          "ChemigationPermitAnnualRecordStatusID": {
            "type": "integer",
            "format": "int32"
          },
          "ChemigationPermitAnnualRecordStatusName": {
            "type": "string",
            "nullable": true
          },
          "ChemigationPermitAnnualRecordStatusDisplayName": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ChemigationPermitAnnualRecordUpsertDto": {
        "required": [
          "ApplicantName",
          "ChemigationInjectionUnitTypeID",
          "ChemigationPermitAnnualRecordStatusID",
          "PivotName",
          "RecordYear"
        ],
        "type": "object",
        "properties": {
          "ChemigationPermitAnnualRecordStatusID": {
            "type": "integer",
            "format": "int32"
          },
          "ApplicantName": {
            "type": "string"
          },
          "PivotName": {
            "type": "string"
          },
          "RecordYear": {
            "type": "integer",
            "format": "int32"
          },
          "DateReceived": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "DatePaid": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "ChemigationInjectionUnitTypeID": {
            "type": "integer",
            "format": "int32"
          },
          "ApplicantPhone": {
            "pattern": "^\\(?([0-9]{3})\\)?[-. ]?([0-9]{3})[-. ]?([0-9]{4})$",
            "type": "string",
            "nullable": true
          },
          "ApplicantMobilePhone": {
            "pattern": "^\\(?([0-9]{3})\\)?[-. ]?([0-9]{3})[-. ]?([0-9]{4})$",
            "type": "string",
            "nullable": true
          },
          "ApplicantMailingAddress": {
            "type": "string",
            "nullable": true
          },
          "ApplicantEmail": {
            "type": "string",
            "nullable": true
          },
          "ApplicantCity": {
            "type": "string",
            "nullable": true
          },
          "ApplicantState": {
            "type": "string",
            "nullable": true
          },
          "ApplicantZipCode": {
            "pattern": "^[0-9]{5}(?:-[0-9]{4})?$",
            "type": "string",
            "nullable": true
          },
          "NDEEAmount": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "ChemicalFormulations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChemigationPermitAnnualRecordChemicalFormulationUpsertDto"
            },
            "nullable": true
          },
          "Applicators": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChemigationPermitAnnualRecordApplicatorUpsertDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ChemigationPermitDto": {
        "type": "object",
        "properties": {
          "ChemigationPermitNumberDisplay": {
            "type": "string",
            "nullable": true
          },
          "ChemigationPermitID": {
            "type": "integer",
            "format": "int32"
          },
          "ChemigationPermitNumber": {
            "type": "integer",
            "format": "int32"
          },
          "ChemigationPermitStatus": {
            "$ref": "#/components/schemas/ChemigationPermitStatusDto"
          },
          "DateCreated": {
            "type": "string",
            "format": "date-time"
          },
          "TownshipRangeSection": {
            "type": "string",
            "nullable": true
          },
          "ChemigationCounty": {
            "$ref": "#/components/schemas/ChemigationCountyDto"
          },
          "Well": {
            "$ref": "#/components/schemas/WellDto"
          }
        },
        "additionalProperties": false
      },
...
      "ChemigationPermitStatusDto": {
        "type": "object",
        "properties": {
          "ChemigationPermitStatusID": {
            "type": "integer",
            "format": "int32"
          },
          "ChemigationPermitStatusName": {
            "type": "string",
            "nullable": true
          },
          "ChemigationPermitStatusDisplayName": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}